GL supersampling for round video, with selectable filter/dither and extended resolution/bitrate
Captures at the sensor's largest near-square SurfaceTexture size under a 1920px
cap (1920x1920 on the Pixel 11 Pro) instead of requesting close to the render
target directly, then downscales to the configured resolution in two separable
GL passes (horizontal + vertical 9-tap kernel, plus ordered dithering in the
vertical pass) rather than the previous single 2x2-tap bilinear resample. The
downscale kernel is user-selectable (Lanczos-2, Box, or Gaussian - A/B testing
found Lanczos clearly better despite its ringing risk) and dither amount is
adjustable including off. Resolution now offers 320/384/448/480/512/640/960px
(ratios against the 1920 capture) and video bitrate 800k-6Mbps.

Bugs found and fixed along the way:
- SCALER_CROP_REGION wasn't set at zoom=1 - harmless at the old small capture
  size, but this sensor's non-square active array (3440x2448) needs an
  explicit crop to produce a square capture at any size, and the HAL's
  undocumented per-resolution default wasn't centered the same way at 1920 as
  it was at the old smaller request. Now always set explicitly, aspect-matched
  to the output, independent of zoom.
- CONTROL_ZOOM_RATIO was never set, and the HAL picked its own default once
  the capture stream got large enough (observed as an unrequested zoom change
  in the HAL log). Now pinned explicitly on every request, kept independent of
  the crop region per CONTROL_ZOOM_RATIO's own documented contract.
- The vertical pass's texture coordinate was derived from the camera-facing
  aTextureCoord attribute, which carries a V-flip/crop convention that has no
  relationship to the plain 2D intermediate texture pass 1 actually wrote -
  this showed up as the frame being shifted inside the circle. Fixed by
  deriving the coordinate from the pass's own clip position instead.
- The ordered-dither shader squared the raw, unwrapped fragment coordinate
  before ever reducing it to the tile size, overflowing mediump/FP16 precision
  above roughly y=256 and clamping a whole band of high-coordinate rows to
  black. Fixed by wrapping to the tile size before the squaring.
- VideoEditedInfo.resultWidth/resultHeight were hardcoded to the legacy 360
  round-video size in three places, regardless of configured resolution -
  TL_documentAttributeVideo.w/h come straight from these fields. Now set from
  the actual encoded videoWidth/videoHeight.
- round_blur_stage_2_frag.glsl (stock's own vignette/blur compositing shader,
  unmodified until now) reads raw pixel coordinates and squares a difference
  inside length() at mediump precision - a real, demonstrable overflow above
  ~512px frame width, though not confirmed as the cause of the black-rim/
  square-fallback symptom investigated this session (stock Android and
  Telegram Web don't show it on the same sent file). Bumped to highp as a
  defensive fix regardless, since it's free and the overflow is real in
  principle even if unconfirmed as this bug's actual cause.

Also recorded: a byte-for-byte confirmation that Telegram does not re-encode
round video server-side (at least not below whatever cap, if any, sits above
the bitrates tested here), and that no mainstream client (stock, Telegram X,
Cherrygram) has ever shipped round video above 512px - the practical safe
ceiling recommended for cross-client playback given the breakage observed
above that on both Android and iOS clients this session.

See FINDINGS.md for the full investigation, including what was ruled out and
what remains genuinely unresolved (the black-rim/square-fallback root cause).

+++ TMessagesProj/src/main/res/raw/round_blur_stage_2_frag.glsl
@@ -1,4 +1,13 @@
-precision mediump float;
+// Stock stayed safe on mediump for a decade because no round video anywhere ever exceeded
+// ~512px (see FINDINGS.md's cross-client survey) - this shader operates on raw gl_FragCoord
+// pixel coordinates (unlike stages 0/1/3, which only touch normalized [0,1] texture coordinates
+// and are resolution-magnitude-independent), and length(center - gl_FragCoord.xy) squares that
+// coordinate internally. At mediump/FP16's ~65504 max representable value, that overflows once
+// |center - fragCoord| exceeds ~256 in either axis - i.e. once the frame is wider than ~512px -
+// producing infinity/NaN that corrupts the sharp/blur mix into black for every fragment beyond
+// that radius. highp has enough range (>3e38) that this never comes close to overflowing at any
+// resolution this app will ever produce.
+precision highp float;
 varying vec2 vTextureCoord;
 
 uniform sampler2D sTexture; // normal texture
